requirements-dev.txt
pytest
requests
build
twine
typing_extensions
pyyaml
mkdocs
mkdocs-material
pymdown-extensions